+++ Tests/Invoke.test.ps1
@@ -10,19 +10,19 @@ task Basic {
 		equals "$r" '{"_id":1,"p1":1} {"_id":2,"p1":"2"} {"_id":3,"p1":3.0}'
 
 		# DELETE with one argument
-		$r = Invoke-LiteCommand 'DELETE test WHERE $._id > @0' 1
+		$r = Invoke-LiteCommand 'DELETE test WHERE _id > @0' 1
 		equals $r 2
 		$r = Invoke-LiteCommand 'SELECT $ FROM test'
 		equals "$r" '{"_id":1,"p1":1}'
 
 		# UPDATE with many arguments
-		$r = Invoke-LiteCommand 'UPDATE test SET p1 = @1, p2 = @2 WHERE $._id = @0' 1, 2, 2
+		$r = Invoke-LiteCommand 'UPDATE test SET p1 = @1, p2 = @2 WHERE _id = @0' 1, 2, 2
 		equals $r 1
 		$r = Invoke-LiteCommand 'SELECT $ FROM test'
 		equals "$r" '{"_id":1,"p1":2,"p2":2}'
 
 		# UPDATE with named parameters
-		$r = Invoke-LiteCommand 'UPDATE test SET p1 = @p1, p2 = @p2 WHERE $._id = @id' @{id = 1; p1 = 3; p2 = 3}
+		$r = Invoke-LiteCommand 'UPDATE test SET p1 = @p1, p2 = @p2 WHERE _id = @id' @{id = 1; p1 = 3; p2 = 3}
 		equals $r 1
 		$r = Invoke-LiteCommand 'SELECT $ FROM test'
 		equals "$r" '{"_id":1,"p1":3,"p2":3}'